Part-way through Component conversion
Started off using MVC for the whole app. Decided to split it up using
entities. Lots of interfaces, and should be able to plug in any
framework (just using libgdx Sprites and Animation classes in the
backend, but pass in RenderComponent so it's polymorphic).

Revised how I was handling animations, and now using a FSM. Works much
better, and is much easier to understand/update.

Backend is using my own InputListener now. I create an AABB based on the
sprite size and implemented FILO in the EventManager.

I still plan to use MVC for the screens, probably using the EventManager
to talk to the controllers and vice versa.