Added a new saveManager class
Interfaces are awesome!

I have a baseSaver class which implements ISaveManager (which has a save
and load method). This baseSave class is abstract and has encrypt and
decrypt save methods. It takes an ISaveEncryption as its only
constructor argument.

The ISaveEncryption interface has both a encryptSave and decrpytSave
method. Both take strings. (All saves are managed as Strings so far).
Any class that implements ISaveEncryption can be passed to a class that
extends BaseSave.

This makes it really easy to update either the way we save, the way we
encrypt, or both.

A JsonSaver and SomeEncryptor class have been created as examples. They
are only examples at the moment to show the process.