AnimationState and new Controller update
Changed how we create AnimationStates. Added an abstract base class
which has an IAnimationComponent. So we now aren't setting the state
every frame, as we just do it in the base constructor.

Implemented the new Controllers mostly. Added a new interface (MyScreen)
to the LibgdxScreens to make my controllers framework agnostic. We pass
a MyScreen through the EventManager to the ScreenListeners
(Controllers), and update it via the interfaces.

Added an ObjectId class (pseudo-enum using static final int fields)
Have to update the ResourceLoader to reflect this change.

Probably a few more little tweaks I missed.