CollisionComponent and UI updates
Refactored how collisions were handled. A CollisionComponent is passed
as an argument to the PhysicsComponent constructor, and it tests it's
object against all the other objects in the world. It holds a list of
objects collided for GameObject-specific PhyicsComponents to use.

Eg. PoopaScoopa uses this list in its PhysicsComponents update methods.
It tests if any of the collided objects are poops, and then removes them
from the world by setting it's clicked state to true.

UI Update:

ShopWindow - updated the way I create the window. An image is used as
the "window", and added to the uiGroup. A table is made and positioned
inside the image then added to the group. Doing it this way allows for
fancier menus.

ItemButton now extends Table, and the table has the clickListener. Can
now add "button" highlighting, add item text, items owned etc, all
neatly packaged in one spot.