BugFixes, and basic debugging
BugFix: Added a check to the Linkmon Wave animation so it ends if the
Linkmon is moving.

Bugfix: Updated LinkmonPhysicsComponent and PoopaScoopa to call
super.update() isntead of both testing for collisions and moving
themselves. Linkmon didn't even have the move test, so it would never
walk around.

Implemented some basic debugging. GameObjects in world, FPS, and Linkmon
Position so far.