Lots of refactoring
Lots of refactoring. We now have a WorldController and a
LibgdxWorldRenderer. WorldRenderer was renamed "UIRenderer".

ControllerServices update method now calls WorldController update.

LibgdxWorldRenderer takes in the world as an argument for renderering
purposes.

GameUi light now functions properly. Removed the redundant (if not
supid) checks for if the button is checked. The event now sends a
boolean based on the button checkedState (isChecked()).

Added a new screen MenuScreen. I will be shifting some of the buttons
from the GameUi screen to the menu. Without a menu screen, I'm pretty
limited by space.

Started building a test Controller class for mini-games. I want to see
if I can create a new world and linkmon, adding user controls for the
linkmon.