Update: Added Item Actions
Rather than having a component for each item, I've decided to use
actions instead. I've created an IItemAction interface which has the
method use(GameObject item, GameObject linkmon, World world) on it.

PoopaScoopa is now able to be placed in the world. Feeding works the
same. Added an isDead boolean to LinkmonStatusComponent, and Revive now
sets isDead() to false.