Online Battling, MiniGames, UI edits
Have been working on the network side of things a bit. The network
controller now handles most of the network events. It has an
OnlineBattle class which has a player and opponent BattleLinkmon. The
network layer sends updates to the network controller and we update the
screen from that.

Note: No UI updates can happen from the network threads. Everything will
have to have a flag or a queue and be checked during an update(); (maybe
send data event to the controller, then another event to the screen to
update next frame by setting a bool)

Added a miniGameController and started looking at using the model to
represent the miniGame. Works pretty well by creating a new World class
and adding objects like before. Added some simple physics to the egg
object component for testing.

Started working on a CyroGenics screen. Just a few buttons and a linkmon
animation so far really.