Refactored how darkening of the world occurs
Darkening of the world was being done in the gameUi. Moved most of the
code to  WorldRenderingComponent class (model) and just test if the
light is on to see if we draw the texture.

Since we now needed 2 draw calls, I had to extend
LibgdxRenderingComponent and add a drawLight() draw method. It has to be
called after all the objects in the world are drawn (WorldRenderer)