Server update, Refactoring of view
Started implementing server-side battle mechanics (I should probably
create a repositry for the server too).

Updated the onlineBattleScreen to display messages based on the update
packet. Need to come up with a way to update health and progress bars
after each message though. Currently they get updated as soon as the
first message appears.

Refactored itemButton widget to a SelectionTable widget. It creates
ISelectables (itemButton implements ISelectable). ISelectable only needs
a reference to the SelectionTable to setSelected now. Previously I was
passing in the screen and setting it via a method call to the screen.

Doing all this lets me reuse the table for other selectable lists I may
wish to use later on. Deciding to have swappable moves prompted this
thought.

Working on cleaning up the views, as I was starting to struggle with how
everything was connected. ItemWindow is a glowing example of how I want
to layout the code.