Major Update
Re-implemented the battle system. Most of the work happens on the
MovePhysicsComponent

Added a sound controller and implemented sounds

In the process of reworking who sends events and if it should just have
a controller reference (View)

Added a flag to the world - isActive. Used for disabling rendering of
the world. I disable the world in the UI menu currently.