Settings, daily rewards, and UI update
Added a settings class and controller that saves a json file for the
settings. Only settings saved are Sound volumes.

Added a daily rewards check to GameUi and made a RewardsWidget using the
LocalMessageBox class. Player stores the last time we received a reward.
Still need to add a variable for the day, and to check against it.
Overall, we will check it has been 16hrs since last reward, then check
it isn't the same day. That should protect against calendar
manipulation.
We are now saving TrainingPoints (formally LinkmonStatPoints), and Daily
Rewards adds 10 - (mistakes*2)

Lots of UI tweaks. Increaed the size of LocalMessageBox, and reskinned.
Added scrolling background to Items Screen.
Changed how SelectionTable was setting "checked" appearance. Rather than
a new texture, just change .setColor(r,g,b,a).
Added itemInfo Label to Items and Shop screens.
Added a tap particle to LocalChatMessage. It was unclear how to progress
the chat previously.
Set Training Screen "Apply" button to hide when no points have been
spent.
Changed the Points Left Label to 2 Labels, on for "Points Left:" and the
other for the actual points variable.